requirements_windows.txt
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.0
fvcore==0.1.5.post20221221
iopath==0.1.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
narwhals==1.40.0
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
plotly==6.1.1
portalocker==3.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32==310
PyYAML==6.0.2
setuptools==80.8.0
six==1.17.0
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
yacs==0.1.8
